+++ Application/Treiber/src/LedTreiber_local.h
@@ -4,7 +4,7 @@
 #include <inttypes.h>
 #include "GlobalLed.h"
 
-void LedTreiber_LedSchalten(uint8_t ledID, LedTreiber_LedFarbe_t ledFarbe);
+void LedTreiber_LedSchalten(uint16_t ledID, LedTreiber_LedFarbe_t ledFarbe);
 
 
 #endif
